requirements.txt
aws-cdk-lib==2.100.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
botocore>=1.29.0
pydantic>=2.0.0
pytest>=7.0.0
moto>=4.0.0